ai-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai-whisper
pydub
ollama
ffmpeg-python
python-multipart